requirements.txt
mysql-connector-python>=9.0.0
python-dotenv>=1.0.1
pytest>=8.3.0
